Move mock server into jest folder
