+++ present_ideas/present_ideas/static/store/__tests__/store.test.ts
@@ -1,5 +1,5 @@
 import { rest } from "msw";
-import { server } from "../../../../../mock_server";
+import { server } from "../../../../../jest/mock_server";
 import { getVueAndStore } from "../../pages/__tests__/utils";
 
 describe("getters", () => {
